api/requirements.txt
Flask
Flask-RESTful
Flask-CORS
python-dotenv